+++ nutanix/resource_nutanix_subnet_test.go
@@ -108,11 +108,7 @@ resource "nutanix_subnet" "acctest-managed" {
 
 func testAccNutanixSubnetConfigUpdate(r int) string {
 	return fmt.Sprintf(`
-data "nutanix_clusters" "clusters" {
-  metadata = {
-    length = 2
-  }
-}
+data "nutanix_clusters" "clusters" {}
 
 output "cluster" {
   value = "${data.nutanix_clusters.clusters.entities.0.metadata.uuid}"
